+++ pdbe/__init__.py
@@ -2,6 +2,6 @@
 Init.
 """
 # pylint:disable=no-name-in-module,import-error
-from pdbe.pdbe import handle_commits_log, handle_checkout, handle_commit_state, put_import_pdb, remove_import_pdb
+# from pdbe.pdbe import handle_commits_log, handle_checkout, handle_commit_state, put_import_pdb, remove_import_pdb
 
 __version__ = '0.2.0'
