+++ pdbe/cli.py
@@ -7,7 +7,16 @@ from os import getcwd, listdir, walk
 from os.path import isfile, join
 from typing import List, Optional, Tuple
 
-from pdbe import handle_commits_log, handle_checkout, handle_commit_state, put_import_pdb, remove_import_pdb
+from commits import (
+    handle_commits_log,
+    handle_checkout,
+    handle_commit_state,
+)
+from pdbe import (
+    put_import_pdb,
+    remove_import_pdb,
+)
+
 
 
 def make_file_state(file_path, clear) -> None:
